requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-generativeai
chromadb
pypdf
lxml
tqdm